+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/UserCardListActionHandler.kt
@@ -3,10 +3,10 @@ package net.pantasystem.milktea.user
 import android.app.Activity
 import net.pantasystem.milktea.common_android.ui.Activities
 import net.pantasystem.milktea.common_android.ui.putActivity
-import net.pantasystem.milktea.user.activity.FollowFollowerActivity
-import net.pantasystem.milktea.user.activity.UserDetailActivity
+import net.pantasystem.milktea.user.followlist.FollowFollowerActivity
 import net.pantasystem.milktea.user.compose.UserDetailCardAction
 import net.pantasystem.milktea.user.compose.UserDetailCardListAction
+import net.pantasystem.milktea.user.profile.UserDetailActivity
 import net.pantasystem.milktea.user.viewmodel.ToggleFollowViewModel
 
 class UserCardListActionHandler(
