+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/search/SearchAndSelectUserScreen.kt
@@ -18,7 +18,6 @@ import androidx.compose.ui.unit.sp
 import kotlinx.coroutines.launch
 import net.pantasystem.milktea.user.R
 import net.pantasystem.milktea.user.compose.SimpleUserListView
-import net.pantasystem.milktea.user.viewmodel.SelectedUserViewModel
 
 @Composable
 @OptIn(ExperimentalMaterialApi::class)
