+++ modules/features/user/src/main/java/net/pantasystem/milktea/user/search/SelectedUserViewModel.kt
@@ -1,4 +1,4 @@
-package net.pantasystem.milktea.user.viewmodel
+package net.pantasystem.milktea.user.search
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
